+++ test-driver-interface-client/src/main/java/de/gematik/test/erezept/remotefdv/client/requests/GetPrescriptionById.java
@@ -21,11 +21,11 @@
 package de.gematik.test.erezept.remotefdv.client.requests;
 
 import com.fasterxml.jackson.core.type.TypeReference;
+import de.gematik.erezept.remotefdv.api.model.Prescription;
 import de.gematik.test.erezept.remotefdv.client.HttpRequestInfo;
 import java.util.List;
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
-import de.gematik.erezept.remotefdv.api.model.Prescription;
 
 @RequiredArgsConstructor
 public class GetPrescriptionById implements PatientRequests<Prescription> {