+++ test-driver-interface-server/src/test/java/de/gematik/test/erezept/remotefdv/server/webservice/mapping/PrescriptionDataMapperTest.java
@@ -20,36 +20,34 @@
 
 package de.gematik.test.erezept.remotefdv.server.webservice.mapping;
 
-import static org.junit.jupiter.api.Assertions.*;
-import static org.mockito.Mockito.mock;
-import static org.mockito.Mockito.when;
-
-import de.gematik.bbriccs.utils.PrivateConstructorsUtil;
 import de.gematik.test.erezept.fhir.builder.kbv.KbvErpBundleFaker;
-import de.gematik.test.erezept.fhir.resources.erp.ErxTask;
-import de.gematik.test.erezept.fhir.resources.kbv.KbvErpMedication;
+import de.gematik.test.erezept.fhir.parser.FhirParser;
+import de.gematik.test.erezept.fhir.r4.erp.ErxTask;
+import de.gematik.test.erezept.fhir.r4.kbv.KbvErpMedication;
 import de.gematik.test.erezept.fhir.values.AccessCode;
 import de.gematik.test.erezept.fhir.values.PrescriptionId;
 import de.gematik.test.erezept.fhir.valuesets.MedicationType;
 import de.gematik.test.erezept.fhir.valuesets.PrescriptionFlowType;
 import de.gematik.test.erezept.remotefdv.server.mapping.PrescriptionDataMapper;
-import java.util.Date;
-import java.util.Optional;
-import java.util.UUID;
 import lombok.val;
 import org.hl7.fhir.r4.model.ResourceType;
 import org.hl7.fhir.r4.model.Task;
 import org.junit.jupiter.api.Test;
 
+import java.util.Date;
+import java.util.Optional;
+import java.util.UUID;
+
+import static org.junit.jupiter.api.Assertions.*;
+import static org.mockito.Mockito.mock;
+import static org.mockito.Mockito.when;
+
 class PrescriptionDataMapperTest {
-  @Test
-  void shouldNotInstantiate() {
-    assertTrue(PrivateConstructorsUtil.isUtilityConstructor(PrescriptionDataMapper.class));
-  }
 
   @Test
   void shouldCreatePrescriptionFromErxTask() {
     val erxTask = mock(ErxTask.class);
+    val fhirParser = new FhirParser();
     val medication = mock(KbvErpMedication.class);
 
     when(medication.getMedicationType()).thenReturn(Optional.of(MedicationType.INGREDIENT));